Misc cleanup and re-org mostly of spectrum code
